+++ packages/vexide-devices/src/smart/motor.rs
@@ -144,8 +144,10 @@ pub enum MotorOptions {
 }
 
 impl Motor {
-    /// The maximum voltage value that can be sent to a [`Motor`].
-    pub const MAX_VOLTAGE: f64 = 12.0;
+    /// The maximum voltage value that can be sent to a V5 [`Motor`].
+    pub const MAX_VOLTAGE_V5: f64 = 12.0;
+    /// The maximum voltage value that can be sent to a EXP [`Motor`].
+    pub const MAX_VOLTAGE_EXP: f64 = 10.0;
 
     /// The rate at which data can be read from a [`Motor`].
     pub const DATA_READ_INTERVAL: Duration = Duration::from_millis(10);
@@ -514,6 +516,14 @@ impl Motor {
             false => Direction::Forward,
             true => Direction::Reverse,
         })
+    }
+
+    /// Gets the maximum voltage for the motor based off of its [motor type](Motor::motor_type).
+    pub const fn max_voltage(&self) -> f64 {
+        match self.motor_type {
+            MotorType::Exp => Self::MAX_VOLTAGE_EXP,
+            MotorType::V5 => Self::MAX_VOLTAGE_V5,
+        }
     }
 
     /// Adjusts the internal tuning constants of the motor when using velocity control.